requirements.txt
tkiniter
newsapi
pillow
urllib
webbrowser
threading